refactor: moves module types and utilities to dedicated file
Centralizes module-related functionality by moving ModuleData interface to global type definitions and extracting route module utilities to a separate module file

Improves code organization and type consistency across components
