+++ src/lib/module.ts
@@ -1,0 +1,3 @@
+export const routeModules = import.meta.glob("/src/routes/**/+page.svelte", {
+	eager: true
+}) as Record<string, App.ModuleData>;
